Rename default to index
