+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as Link from "gatsby-link";
 import HeaderMenu from "../components/HeaderMenu/HeaderMenu";
-import { menuItems } from "../layouts/default";
+import { menuItems } from "../layouts";
 import {
   Button,
   Segment,
